fix(jobs): block IPv4-mapped/embedded IPv6 webhook SSRF targets
is_disallowed_webhook_ipv6 only checked native IPv6 ranges, so the
cloud-metadata IP and loopback/private ranges were reachable via the
IPv6 literal form (::ffff:169.254.169.254, ::127.0.0.1, 6to4/Teredo).
Decode any embedded IPv4 and run the v4 filter on it.
